make several functions for testing compartments
